move security into one place
